backend/requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai
numpy
    